refactor: move all pages to a folder
